backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.0.2
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
pytz==2023.3